+++ controller/getAllUserPost.js
@@ -1,0 +1,33 @@
+const UserModel = require("../model/userSchema");
+
+// Get all post by user
+exports.getAllUserPost = async (req, res) => {
+  try {
+    // Deconstruct req
+    const {
+      params: { username },
+      query: { page, limit },
+    } = req;
+
+    // Store to variable
+    const data = await UserModel.find({ username: username }).populate({
+      path: "posts",
+      limit: limit,
+      skip: (page - 1) * limit,
+      sort: { date: 1 },
+    });
+
+    //
+
+    res.status(200).json({
+      message: "success",
+      length: data[0].posts.length,
+      page: page,
+      data: data[0].posts,
+    });
+  } catch (err) {
+    res.status(400).json({
+      message: err.message,
+    });
+  }
+};
