+++ controller/ping.js
@@ -1,0 +1,5 @@
+exports.ping = (req, res) => {
+  res.status(201).json({
+    message: "Ping",
+  });
+};
