+++ model/userSchema.js
@@ -4,21 +4,28 @@ const UserSchema = new mongoose.Schema({
   username: {
     type: String,
     unique: true,
+    required: true,
   },
   password: {
     type: String,
+    required: true,
   },
   firstName: {
     type: String,
+    required: true,
   },
   lastName: {
     type: String,
+    required: true,
   },
+
   profilePic: {
-    type: String,
+    image: Buffer,
+    contentType: String,
   },
   cover: {
-    type: String,
+    image: Buffer,
+    contentType: String,
   },
   friends: [
     {
