+++ package.json
@@ -15,6 +15,8 @@
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^7.3.1",
-    "nodemon": "^2.0.22"
+    "multer": "^1.4.5-lts.1",
+    "nodemon": "^2.0.22",
+    "paginate": "^0.2.0"
   }
 }
